Specify priority & scheduling for systemd service
OOMScoreAdjust is used to disable this service from OOM killer consideration, as this will explicitly be recovering memory.

IO and CPU scheduling (incl. niceness) both reflect the lowest priorities possible, as keeping the data in tmpfs longer helps performance anyway.

In the future, it may be better to dynamically choose the scheduling policies used, depending on memory pressure or integrity requirements.